+++ packages/react-dashboard/src/DashboardEditToolbar.tsx
@@ -100,31 +100,39 @@ export const DashboardEditToolbar = () => {
 
   return (
     <>
-      <Flex sx={{ gap: '3', alignItems: 'center', flexWrap: 'wrap' }}>
+      <Flex
+        sx={{
+          gap: '3',
+          flexWrap: 'wrap',
+        }}
+      >
         <Button
-          variant="accent"
+          variant="secondary"
           onClick={() => {
             setAddCardDrawerOpen(true);
           }}
           rightIcon="lucide:plus"
-          sx={{ marginRight: '10' }}
         >
-          Adicionar item
+          Adicionar Métricas
         </Button>
         {selectedTemplate?.editable && (
-          <Button variant="accent" onClick={saveEdit} rightIcon="lucide:save">
+          <Button
+            variant="secondary"
+            onClick={saveEdit}
+            rightIcon="lucide:save"
+          >
             Salvar
           </Button>
         )}
         <Button
-          variant="primary"
+          variant="secondary"
           onClick={saveAsNew}
           rightIcon="lucide:copy-check"
         >
-          Salvar como novo
+          Salvar Novo Template
         </Button>
         <Button variant="secondary" onClick={cancelEdit} rightIcon="lucide:x">
-          Cancelar edição
+          Cancelar Edição
         </Button>
       </Flex>
 
@@ -137,8 +145,8 @@ export const DashboardEditToolbar = () => {
         direction="right"
         size={drawerSizeXs}
         sx={{
-          width: addCardDrawerOpen ? '100%' : '0',
-          display: addCardDrawerOpen ? 'block' : 'none',
+          position: 'fixed',
+          pointerEvents: addCardDrawerOpen ? 'auto' : 'none',
         }}
       >
         <Flex
@@ -157,10 +165,10 @@ export const DashboardEditToolbar = () => {
             }}
           >
             <Text sx={{ fontWeight: 'bold', fontSize: 'lg' }}>
-              Adicionar item
+              Adicionar métricas
             </Text>
             <Text sx={{ color: 'text.muted', fontSize: 'sm' }}>
-              Escolha o tipo de card para adicionar ao dashboard.
+              Escolha a métrica para adicionar ao dashboard.
             </Text>
             <Search
               key={addCardDrawerOpen ? 'open' : 'closed'}
@@ -255,8 +263,8 @@ export const DashboardEditToolbar = () => {
         direction="right"
         size={drawerSizeSm}
         sx={{
-          width: saveAsNewModalOpen ? '100%' : '0',
-          display: saveAsNewModalOpen ? 'block' : 'none',
+          position: 'fixed',
+          pointerEvents: saveAsNewModalOpen ? 'auto' : 'none',
         }}
       >
         <Flex