+++ docs/storybook/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.24.1](https://github.com/ttoss/ttoss/compare/@docs/storybook@3.24.0...@docs/storybook@3.24.1) (2026-04-17)
+
+**Note:** Version bump only for package @docs/storybook
+
+# [3.24.0](https://github.com/ttoss/ttoss/compare/@docs/storybook@3.23.0...@docs/storybook@3.24.0) (2026-04-17)
+
+### Features
+
+- **react-dashboard:** add onEditingGridChange, card utils, period presets, currency prop ([#947](https://github.com/ttoss/ttoss/issues/947)) ([8c4d4a2](https://github.com/ttoss/ttoss/commit/8c4d4a206956fd6a2237055dbae7036db1324d9d))
+
 # [3.23.0](https://github.com/ttoss/ttoss/compare/@docs/storybook@3.22.5...@docs/storybook@3.23.0) (2026-04-14)
 
 ### Features