+++ terezinha-farm/vite-app/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.17.25](https://github.com/ttoss/ttoss/compare/@terezinha-farm/vite-app@0.17.24...@terezinha-farm/vite-app@0.17.25) (2026-04-17)
+
+**Note:** Version bump only for package @terezinha-farm/vite-app
+
+## [0.17.24](https://github.com/ttoss/ttoss/compare/@terezinha-farm/vite-app@0.17.23...@terezinha-farm/vite-app@0.17.24) (2026-04-17)
+
+**Note:** Version bump only for package @terezinha-farm/vite-app
+
 ## [0.17.23](https://github.com/ttoss/ttoss/compare/@terezinha-farm/vite-app@0.17.22...@terezinha-farm/vite-app@0.17.23) (2026-04-09)
 
 **Note:** Version bump only for package @terezinha-farm/vite-app